externref

Attribute Macro externref

#[externref]

Available on crate feature macro only.

Prepares imported functions or an exported function with Resource args and/or return type.

Inputs

This attribute must be placed on an extern "C" { ... } block or an extern "C" fn. If placed on block, all enclosed functions with Resource args / return type will be wrapped.

Processing

The following arg / return types are recognized as resources:

Resource<..>, &Resource<..>, &mut Resource<..> with one or two type args

ResourceCopy<..> and its references

Option<_> of the above six variations

For complex cases, resource detection can be controlled with a #[resource] attribute. This attribute can be placed on a function arg or on the function itself (in which case it corresponds to the return type; attributes cannot be placed on the return type directly).

#[resource], #[resource = true] or #[resource(true)] mark an arg / return type as a resource.

#[resource = false] or #[resource(false)] mark an arg / return type as a non-resource.