externref::processor

Enum Error

#[non_exhaustive]
pub enum Error {
    Read(ReadError),
    Wasm(Error),
    UnexpectedImportType {
        module: String,
        name: String,
    },
    NoExport(String),
    UnexpectedExportType(String),
    UnexpectedArity {
        module: Option<String>,
        name: String,
        expected_arity: usize,
        real_arity: usize,
    },
    UnexpectedType {
        module: Option<String>,
        name: String,
        location: Location,
        real_type: ValType,
    },
    IncorrectGuard {
        function_name: Option<String>,
        code_offset: Option<u32>,
    },
    UnexpectedCall {
        function_name: Option<String>,
        code_offset: Option<u32>,
    },
}

Available on crate feature processor only.

Errors that can occur when processing a WASM module.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Read(ReadError)

Error reading the custom section with function declarations from the module.

Wasm(Error)

Error parsing the WASM module.

UnexpectedImportType

Unexpected type of an import (expected a function).

Fields

module: String

Name of the module.

name: String

Name of the function.

NoExport(String)

Missing exported function with the enclosed name.

UnexpectedExportType(String)

Unexpected type of an export (expected a function).

UnexpectedArity

Imported or exported function has unexpected arity.

Fields

module: Option<String>

Name of the module; None for exported functions.

name: String

Name of the function.

expected_arity: usize

Expected arity of the function.

real_arity: usize

Actual arity of the function.

UnexpectedType

Argument or return type of a function has unexpected type.

Fields

module: Option<String>

Name of the module; None for exported functions.

name: String

Name of the function.

location: Location

Location of an argument / return type in the function.

real_type: ValType

Actual type of the function (the expected type is always i32).

IncorrectGuard

Incorrectly placed externref guard. This is caused by processing the WASM module with external tools (e.g., wasm-opt) before using this processor.

Fields

function_name: Option<String>

Name of the function with an incorrectly placed guard.

code_offset: Option<u32>

WASM bytecode offset of the offending guard.

UnexpectedCall

Unexpected call to a function returning externref. Such calls should be confined in order for the processor to work properly. Like with Self::IncorrectGuard, such errors should only be caused by external tools (e.g., wasm-opt).

Fields

function_name: Option<String>

Name of the function containing an unexpected call.

code_offset: Option<u32>

WASM bytecode offset of the offending call.