externref

Trait DropGuard

pub trait DropGuard: Sealed { }

Trait for various drop behaviors for Resources. This constrains the second type param in Resource.

The contents of this trait is an implementation detail. It cannot be implemented for external types.

Currently, 2 implementations are available:

Register is the default implementation that implements RAII-style cleanup on drop, including calling a customizable hook if one was supplied to the Processor.

Forget is a no-op implementation corresponding to ResourceCopy.

See Resource and ResourceCopy docs for more context and examples of usage.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.