externref

Type Alias ResourceCopy

pub type ResourceCopy<T> = Resource<T, Forget>;

Resource variation that can be copied.

Cleanup

ResourceCopy does not clean up the externref table entry on drop. It can only be cleaned up by the host side, or by implementing custom Drop logic for a higher-level Resource wrapper. In the extreme case, when the WASM module is short-lived, garbage collection of dead externrefs may be summarily ignored.

For custom Drop logic, it may be useful to pass ResourceCopy<_> by value as a non-resource (see the externref macro docs) as follows.

use externref::{externref, ResourceCopy};

#[externref]
// ^ In this particular case, this attribute may be skipped.
#[link(wasm_import_module = "data")]
unsafe extern "C" {
    fn custom_drop(#[resource = false] data: ResourceCopy<CustomDrop>);
    // The host will receive `data: usize` - the 0-based index
    // of `externref` table entry the resource points to.
}

struct CustomDrop(ResourceCopy<Self>);

impl Drop for CustomDrop {
    fn drop(&mut self) {
        unsafe { custom_drop(self.0); }
    }
}

This relies on the fact that ResourceCopy<_> is guaranteed to have the identical layout as usize.

Aliased Type

pub struct ResourceCopy<T> { /* private fields */ }